Move all test file into one folder and change the layout for home page to original design
