+++ src/__test__/CoinCard.test.js
@@ -3,8 +3,8 @@ import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from '../../redux/configureStore';
-import CoinCard from '../CoinCard';
+import store from '../redux/configureStore';
+import CoinCard from '../components/CoinCard';
 
 describe('Testing CoinCard component', () => {
   const coin = {
